+++ CHANGELOG.md
@@ -10,7 +10,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 - Switch build tools to [pixi].
-- Upgraded [zensols.deeplearn] to 1.14.2.
+- Removed [zensols.deeplearn] dependency.
+- Upgrade dependencies: `torch` `transformers`, `datasets`, `accelerate`,
+  `peft`, `trl`.
+- Removed default quantization configuration.
+- Add LoRA generator configuration.
+- Separate and configure LoRA adapter output directory.
+- Add trainer evaluation dataset.
 
 
 ## [0.0.1] - 2025-05-04